[PVR] Refactor CPVRGUIActions step 1: Move timer related actions to own class.
